+++ src/volumito/__init__.py
@@ -17,7 +17,7 @@ from volumito.clients import (
     VolumioRESTAPIClient,
 )
 
-__version__ = "0.0.25"
+__version__ = "0.0.26"
 __author__ = "Alberto Pettarin"
 __email__ = "alberto@albertopettarin.it"