+++ docs/CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.0.26] - 2026-08-03
+
+### Changed
+
+- Renamed the `queue get` command to `queue list` and the corresponding configuration key
+- Renamed the `zones get` command to `zones list` and the corresponding configuration key
+
+
 ## [0.0.25] - 2026-08-03
 
 ### Added